// Backend/src/models/scriptForm.ts
export interface ScriptForm {
  equipBrand: string;
  equipModel: string;
  switchName: string;
  globalPassword: string;
  adminPassword: string;
  encryptAdmin: string;
  encryptGlobal: string;
  vlan: string;
  ports: string;
  managementVlan: string;
  ip: string;
  mask: string;
  gateway: string;
  banner: string;
  consolePassword: string;
}